+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fix bug with a hardcoded VAT percentage
 - add methods to cast response errors to more informative objects
+- add option to use reverse charge VAT
 
 ### 1.0.1